group_manager screen edit to managers
