+++ src/constants/action-types.js
@@ -17,3 +17,5 @@ export const SET_CONTENTS = 'SET_CONTENTS'
 export const SET_GROUP_MANAGERS = 'SET_GROUP_MANAGERS'
 
 export const SET_USERS = 'SET_USERS'
+
+export const SET_MANAGERS= 'SET_MANAGERS'
