+++ src/pages/Home/components/Managers/services/createManager.js
@@ -17,4 +17,5 @@ const createManagers = async (data, token) =>
       console.log(e);
       return { data: {}, errors: e }
     });
+    
 export default createManagers;
