+++ src/pages/Home/components/Managers/services/deleteManager.js
