+++ src/pages/Home/components/Managers/services/editManager.js
