+++ src/pages/Home/components/Managers/styles.js
